# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Runpod Storage API",
    "description": "\n        Professional REST API for Runpod network storage management.\n\n        This API provides comprehensive access to Runpod storage operations including:\n        - Network volume management (create, list, delete)\n        - File operations (upload, download, list, delete)\n        - Robust error handling and validation\n        - OpenAPI 3.0 compliant documentation\n\n        ## Authentication\n\n        All endpoints require authentication via API key. You can authenticate in several ways:\n\n        1. **Bearer Token Header**: `Authorization: Bearer your-api-key`\n        2. **API Key Header**: `X-API-Key: your-api-key`\n        3. **Query Parameter**: `?api_key=your-api-key`\n\n        Get your API key from [Runpod Console](https://console.runpod.io/user/settings).\n\n        ## Rate Limiting\n\n        API requests are subject to rate limiting. See response headers for current limits.\n\n        ## SDKs and Tools\n\n        - **Python SDK**: `pip install runpod-storage`\n        - **CLI Tool**: `runpod-storage --help`\n        - **Docker Image**: `docker run runpod/storage-api`\n        ",
    "contact": {
      "name": "Runpod Storage Team",
      "url": "https://github.com/runpod/runpod-storage",
      "email": "support@runpod.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
...
  "paths": {
    "/api/v1/volumes": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "List network volumes",
        "description": "Retrieve a list of all network volumes associated with your account.",
        "operationId": "list_volumes_api_v1_volumes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "api_key_header",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Header"
            }
          },
          {
            "name": "api_key_query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListVolumesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "Authorization failed"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Create network volume",
        "description": "Create a new network volume with specified name, size, and datacenter.",
        "operationId": "create_volume_api_v1_volumes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "api_key_header",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Header"
            }
          },
          {
            "name": "api_key_query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Query"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVolumeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkVolume"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "Authorization failed"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/volumes/{volume_id}": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "Get volume details",
        "description": "Retrieve detailed information about a specific network volume.",
        "operationId": "get_volume_api_v1_volumes__volume_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Volume Id"
            }
          },
          {
            "name": "api_key_header",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Header"
            }
          },
          {
            "name": "api_key_query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkVolume"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "Authorization failed"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Storage"
        ],
        "summary": "Update volume",
        "description": "Update a network volume's name and/or size. Size can only be increased.",
        "operationId": "update_volume_api_v1_volumes__volume_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Volume Id"
            }
          },
          {
            "name": "api_key_header",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Header"
            }
          },
          {
            "name": "api_key_query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Query"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetworkVolumeUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkVolume"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "Authorization failed"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Storage"
        ],
        "summary": "Delete volume",
        "description": "Delete a network volume. This operation is irreversible.",
        "operationId": "delete_volume_api_v1_volumes__volume_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Volume Id"
            }
          },
          {
            "name": "api_key_header",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Header"
            }
          },
          {
            "name": "api_key_query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "Authorization failed"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/volumes/{volume_id}/files/list": {
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "List files in volume",
        "description": "List all files in a network volume, optionally filtered by prefix. Requires S3 credentials in request body.",
        "operationId": "list_files_api_v1_volumes__volume_id__files_list_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Volume Id"
            }
          },
          {
            "name": "api_key_header",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Header"
            }
          },
          {
            "name": "api_key_query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Query"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListFilesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListFilesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "Authorization failed"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/volumes/{volume_id}/files": {
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Upload file",
        "description": "Upload a file to a network volume. Supports large files via multipart upload. Requires S3 credentials in form data.",
        "operationId": "upload_file_api_v1_volumes__volume_id__files_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Volume Id"
            }
          },
          {
            "name": "api_key_header",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Header"
            }
          },
          {
            "name": "api_key_query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Query"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v1_volumes__volume_id__files_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "Authorization failed"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/volumes/{volume_id}/files/download": {
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Download file",
        "description": "Download a file from a network volume. Requires S3 credentials in request body.",
        "operationId": "download_file_api_v1_volumes__volume_id__files_download_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Volume Id"
            }
          },
          {
            "name": "api_key_header",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Header"
            }
          },
          {
            "name": "api_key_query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Query"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "Authorization failed"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/volumes/{volume_id}/files/delete": {
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Delete file",
        "description": "Delete a file from a network volume. Requires S3 credentials in request body.",
        "operationId": "delete_file_api_v1_volumes__volume_id__files_delete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Volume Id"
            }
          },
          {
            "name": "api_key_header",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Header"
            }
          },
          {
            "name": "api_key_query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key Query"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "Authorization failed"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Delete success status"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Status message",
            "example": "Successfully deleted"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "DeleteResponse",
        "description": "Response for delete operations."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListVolumesResponse": {
        "properties": {
          "volumes": {
            "items": {
              "$ref": "#/components/schemas/NetworkVolume"
            },
            "type": "array",
            "title": "Volumes",
            "description": "List of network volumes"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total number of volumes"
          }
        },
        "type": "object",
        "required": [
          "volumes",
          "total_count"
        ],
        "title": "ListVolumesResponse",
        "description": "Response for listing volumes."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}